Add documentation in the code.

+++ app/Controllers/Category.php
@@ -5,11 +5,17 @@ namespace App\Controllers;
 use App\Controllers\BaseController;
 use App\Models\CategoryModel;
 use App\Filters\AuthFilter;
+use App\Models\NewsModel;
 
 class Category extends BaseController
 {
     protected $filters = ['auth'];
 
+    /**
+     * Display the index page for managing categories in the admin area.
+     *  
+     * @return string Rendered HTML content for the admin index page.
+     */
     public function index()
     {
         $categoryModel      = Model(CategoryModel::class);
@@ -20,6 +26,11 @@ class Category extends BaseController
         return parent::renderTemplate($content, $data);
     }
 
+    /**
+     * Display the create category page in the admin area.
+     *  
+     * @return string Rendered HTML content for the create category page.
+     */
     public function create()
     {
         $data['title']            = 'New Category';
@@ -29,6 +40,13 @@ class Category extends BaseController
         return parent::renderTemplate($content, $data);
     }
 
+    /**
+     * Display the edit a especific category page in the admin area.
+     * 
+     * @param int $id A specific id of the category
+     *
+     * @return string Rendered HTML content for the create category page.
+     */
     public function edit($id)
     {
         $categoryModel            = Model(CategoryModel::class);
@@ -56,10 +74,22 @@ class Category extends BaseController
         return $this->response->redirect(site_url('admin/index'));
     }
 
+    /**
+     * Delete a specific category without news relationed.
+     * 
+     * @param int $id A specific id of the category
+     * 
+     * @return \CodeIgniter\HTTP\RedirectResponse|\CodeIgniter\HTTP\Response redirection.
+     */
     public function delete($id = null)
     {
+        $newsModel     = model(NewsModel::class);
         $categoryModel = model(CategoryModel::class);
-        $categoryModel->where('id', $id)->delete();
+
+        $relatedNews = $newsModel->where('category_id', $id)->countAllResults();
+        if ($relatedNews == 0) {
+            $categoryModel->where('id', $id)->delete();
+        }
 
         return $this->response->redirect(site_url('admin/index'));
     }
